+++ test/Config.test.ts
@@ -13,7 +13,7 @@ import {readFileSync, writeFileSync, appendFile, ensureFileSync} from 'fs-extra'
 export class SysTest extends SysNode {
 	protected	override readFileSync = async (path: string)=> readFileSync(path, {encoding: 'utf8'});
 	protected	override writeFileSync = async (path: string, data: Buffer, o?: object)=> writeFileSync(path, data, o);
-	override appendFile = async (path: string, data: string | Buffer, callback: (err: NodeJS.ErrnoException)=> void)=> appendFile(path, data, callback);
+	override appendFile = async (path: string, data: string, callback: (err: NodeJS.ErrnoException)=> void)=> appendFile(path, data).catch(err=> callback(err));
 	override ensureFileSync = async (path: string)=> ensureFileSync(path);
 }
 //===== Test Class =====